refactor(http): refactor http module to enhance imports readability
