+++ app/layout.tsx
@@ -7,7 +7,7 @@ import FooterContainer from '@/app/components/structure/footer'
 import HeaderContainer from '@/app/components/structure/header'
 import ButtonCloseWebview from '@/app/components/ui/button-close-webview'
 import { LANGUAGE, SITE_NAME } from '@/app/constants/config'
-import { HttpProvider } from '@/app/services/http'
+import { HttpProvider } from '@/app/services/http/providers'
 import ILayout from '@/app/types/layout'
 import { NativeBridgeProvider } from '@/app/utils/native/bridge/native-bridge-provider/native-bridge-provider'
 import { DevTokenSetter } from '@/app/utils/native/token/dev-token-setter/dev-token-setter'
